#881 Custom Mapper Lookup Table (#1960)
* tests and stuff

* target

* took me long enough

* humidity sensor now u16 to i16

* clean up

* changed some values

* comments

+++ Parametric/src/Sensor/SensorType.lua
@@ -15,7 +15,7 @@ return function(core)
   end
 
   local unpack_table = function(table)
-    return structure(i32(table.raw), i32(table.mapped))
+    return structure(u16(table.raw), i16(table.mapped))
   end
 
   local generate_table = memoize(function(sensorTable)
